+++ Code/Modules/Xen/EngineConfig.hpp
@@ -5,7 +5,6 @@
 #pragma once
 
 #include <Common/XenCommon.hpp>
-
 #include <ini.h>
 
 namespace Xen {